requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
bcrypt==3.2.0
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
google-generativeai>=0.8.0
google-cloud-speech==2.26.1
google-cloud-texttospeech==2.16.3
psycopg2-binary
passlib[argon2]
argon2-cffi
